Wait for client cache to expire before running tests
The client used will cache objects created/updated and unfortunately the
tests run quicker than the expire time, which means that we'll get stale
entries from the cache back.

The main issue is that the tests aren't 100% deterministic as we might
not get what we expect from the client, so we probably need to be more
vigilant when writing tests and wait for the conditions that we actually
expect before asserting them.