Delete everything in AfterEach hook
With the previous cleanup I've seen resources linger around, and while
I'm not 100% sure why that happens, with this it doesn't seem to /shrug.

I did see that the GitTrackObject controller sometimes received a
Reconcile request while cleanup was taking place in the `AfterEach`
hook, as it's stopped after the cleanup has taken place. For whatever
reason I've not seen that happen with this approach (i.e. listing and
deleting everything).