Update `git-store` to v0.6.0
The new release brings with it support for passing a `context.Context`
to its `Repo.{Checkout,Fetch}` methods (via the new methods
`Repo.{Checkout,Fetch}Context`. This is useful for adding timeouts as
both of these methods perform network calls (for fetching changes from
remote repositories).

It also includes a change that will lower the CPU usage, as the file log
is now computed lazily on demand (and is actually not used in Faros at
all). Should be looking at about a 50% decrease in CPU usage.