Add initial tests for files-related functions
I'm not terribly fond of the tests for `objectsFrom` depending on the
files returned from `getFiles`, but I couldn't find any nice way of
creating static fixtures through the git library we're using.
For now it's fine, but it would be good to decouple those tests from
depending on an actual repository.