Mimic `kubectl apply -f <path>` behaviour in GitTrack controller
Previously if we encountered a file that we could partially parse we
would discard everything in said file, and eventually we would clean up
any "leftover" resources (which could be ones that we for some reason
now couldn't parse, perhaps because someone introduce a syntax error or
similar).

Now we instead try to partially apply what we can parse, which is what
`kubectl apply -f <path>` does as well. Hopefully this should be of
least surprise to users.